ID-498 Filter spammy sentry issues.
